Update the icon design
After the first build and run in debug mode it turned out that the
originally created icon did not match expected layout (size especially).
The designs had to be adjusted so that the icon drawings are occupying
most of the space and look more readable and clear when installed in the
browser.
All elements have been enlarged but the layout itself of elements have
been left intact.
The only change to the project of the icon is to correct the project
regarding some unused elements, layers and incorrect chore minor things.